+++ src/module.rs
@@ -27,7 +27,7 @@ impl<'ctx> Module<'ctx> {
     pub fn run_test(&self) {
         self.module_ir.print_to_stderr();
 
-        type TestFunc = unsafe extern "C" fn(i32, i32, i32, bool) -> i64;
+        type TestFunc = unsafe extern "C" fn(u8, i16, i32, bool) -> i64;
 
         let execution_engine = self
             .module_ir
@@ -38,8 +38,8 @@ impl<'ctx> Module<'ctx> {
             let test_func: JitFunction<'_, TestFunc> =
                 execution_engine.get_function("test").unwrap();
 
-            let x = 1i32;
-            let y = 2i32;
+            let x = 1u8;
+            let y = 2i16;
             let z = 3i32;
             let w = true;
             dbg!(test_func.call(x, y, z, w));
